+++ Programming/C++/Problem4/highestScorer.cpp
