+++ Programming/Golang/Problem3/main.go
