+++ Programming/Java/CountDigitsInO(1)/countDigits.java
