+++ Programming/Java/HelloWorldCustomClassName/AkDemo.java
@@ -1,0 +1,10 @@
+import java.io.*;
+
+class AkDemo
+{
+	public static void main(String args[])
+	{
+                                  // Using my Class Name
+		Akshat.ak.println("Hello World");
+	}
+}
