+++ Programming/Java/HelloWorldCustomClassName/Akshat.java
@@ -1,0 +1,5 @@
+import java.io.*;
+class Akshat {
+static{}
+static PrintWriter ak=new PrintWriter(System.out,true);
+}
