--- Programming/Java/HelloWorld_CustomClassName/AkDemo.java
@@ -1,10 +1,0 @@
-import java.io.*;
-
-class AkDemo
-{
-	public static void main(String args[])
-	{
-                                  // Using my Class Name
-		Akshat.ak.println("Hello World");
-	}
-}
