--- Programming/Java/HelloWorld_CustomClassName/Akshat.java
@@ -1,5 +1,0 @@
-import java.io.*;
-class Akshat {
-static{}
-static PrintWriter ak=new PrintWriter(System.out,true);
-}
