+++ Programming/Java/Problem6/treeProblem1.java
