+++ Programming/Java/Problem7/s3.java
@@ -1,0 +1,23 @@
+static boolean areAnagram(char s1[], char s2[]) 
+    { 
+        int count1[] = new int[NO_OF_CHARS]; 
+        Arrays.fill(count1, 0); 
+        int count2[] = new int[NO_OF_CHARS]; 
+        Arrays.fill(count2, 0); 
+        int i; 
+
+        for (i = 0; i < s1.length && i < s2.length; 
+             i++) { 
+            count1[s1[i]]++; 
+            count2[s2[i]]++; 
+        } 
+  
+        if (s1.length != s2.length) 
+            return false; 
+  
+        for (i = 0; i < NO_OF_CHARS; i++) 
+            if (count1[i] != count2[i]) 
+                return false; 
+  
+        return true; 
+    }
