+++ Programming/Python/Problem6/CountingWordsStartingWithVowels.py
